Implement complete stickers feature - move from task docs to actual implementation with components, hooks, types, and stores
