+++ qcut/apps/web/src/components/editor/media-panel/views/stickers/index.tsx
@@ -1,0 +1,10 @@
+// Main export for stickers module
+export { StickersView } from "./stickers-view";
+
+// Re-export types if needed elsewhere
+export type { 
+  StickerItemProps,
+  CollectionContentProps,
+  StickersViewProps,
+  RecentSticker 
+} from "./types/stickers.types";
